+++ tests/stdlib/tmemlinesBuf.nim
@@ -1,5 +1,5 @@
 import memfiles
-var inp = memfiles.open("readme.txt")
+var inp = memfiles.open("tests/stdlib/tmemlinesBuf.nim")
 var buffer: TaintedString = ""
 for line in lines(inp, buffer):
   echo("#" & line & "#")